Crate atcoder_snippets

Utility snippets for fighting AtCoder.

Modules

bsearch
Generalized binary search.

cmp
Comparation and sorting.

collections
Data structures.

geometry
2D Geometric calculations.

interval
Intervals on a number line.

iter
Enriches iterators.

modulo

num
Numeric types and traits.

option
Enriches bool and Option.

range
Enriches ranges

read
Macros and functions for reading problem input.

rolling_hash

slice
Enriches slices.

table
2-dimentional array.

utils
Utilities.

vec
2D and 3D geometric vectors.

write
Output to stdout.

xorshift
Xorshift random number generator.

z
Exact string matching by Z algorithm.

Macros

chmax
Assigns the given value if it is greater than the current one.

chmin
Assigns the given value if it is smaller than the current one.

dbg
Make a debug output of the given expression to stderr.

echo
Output values by println!("{} {} ... {}", value_1, value_2, ..., value_n)`.

hash_counter

list
Makes a list by enumerating its contents.

read
標準入力から一行を読み込み、結果を変数に代入する。

read_chunk
Reads multiple lines from stdin and create let bindings.

readable
Makes a type readable from stdin.

readable_collection
Make collection type readable from input line.

table
Creates a table in a simlar way to vec! macro.