Trait atcoder_snippets::bsearch::BSearch

pub trait BSearch: Sized {
    type Item;
    fn is_empty(&self) -> bool;

    fn leftmost_item(&self) -> Self::Item;

    fn rightmost_item(&self) -> Self::Item;

    fn middle_item(&self) -> Self::Item;

    fn left_half(&self) -> Self;

    fn right_half(&self) -> Self;

    fn is_bsearch_converged(&self) -> bool;

    fn bsearch_left_max<F>(&self, is_left: F) -> Option<Self::Item>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }

    fn bsearch_right_min<F>(&self, is_right: F) -> Option<Self::Item>
    where
        F: FnMut(&Self::Item) -> bool,
    { ... }
}

A sequence that binary search is applicable to.

Associated Types

type Item

Item type of the sequence.

Loading content...
Required methods

fn is_empty(&self) -> bool

Returns whether the sequence is empty.

fn leftmost_item(&self) -> Self::Item

Returns the last item in the sequence.

If the sequence is empty, this method may panic.

fn rightmost_item(&self) -> Self::Item

Returns the first item in the sequence.

If the sequence is empty, this method may panic.

fn middle_item(&self) -> Self::Item

Returns the middle item in the sequence.

If the sequence is discrete and contains an odd number of items, this method returns the exact middle of the sequence. If the sequence is discrete and contains an even number of items, it's not sure whether this method returns the left middle or the right middle.

For the same two sequeces, this method always returns the same value.

If the sequence is empty, this method may panic.

fn left_half(&self) -> Self

Returns the left half of the sequence, including self.middle_item() placed at the last of the returned sequence.

For any non-empty and non-converged sequence, this method returns an non-empty sequence.

If the sequence is empty or already converged, this method may panic.

fn right_half(&self) -> Self

Returns the right half of the sequence, including self.middle_item() placed at the first of the returned sequence.

For any non-empty and non-converged sequence, this method returns an non-empty sequence.

If the sequence is empty of already, this method may panic.

fn is_bsearch_converged(&self) -> bool

Check convergence of halving process.

An empty sequence returns true.

For discrete sequences, this method is equivalent that len <= 2.

Loading content...
Provided methods

fn bsearch_left_max<F>(&self, is_left: F) -> Option<Self::Item> where
F: FnMut(&Self::Item) -> bool,

Returns the rightmost item satisfing is_left.

When calling bsearch_left_max, you must make sure that the sequence can be partitioned by is_left. That is, there must be a position satisfying that:

All items left of the position satisfies is_left.

All items right of the position doesn't satisfy is_left.

The situation can be divided the following 4 cases:

The sequence is empty. Returns None.

The sequence is partitioned into non-empty left and right parts: [L, L, L, L, R, R]. Returns Some(the rightmost L item).

All items satisfy is_left: [L, L, L, L, L, L]. Returns Some(the rightmost item).

All items don't satisfy is_left: [R, R, R, R, R, R]. Returns None.

Example

let seq = [3, 6, 9, 12, 15];
assert_eq!((0..seq.len()).bsearch_left_max(|&i| seq[i] <= 10), Some(2));

fn bsearch_right_min<F>(&self, is_right: F) -> Option<Self::Item> where
F: FnMut(&Self::Item) -> bool,

Returns the minimum item satisfing is_right.

See doc of bsearch_left_max.

Example

let seq = [3, 6, 9, 12, 15];
assert_eq!((0..seq.len()).bsearch_right_min(|&i| seq[i] >= 10), Some(3));

Loading content...