Trait atcoder_snippets::cmp::IteratorMinmax

pub trait IteratorMinmax: Iterator {
    fn minmax(self) -> Option<(Self::Item, Self::Item)>;

    fn minmax_by_key<K, F>(self, key_fn: F) -> Option<(Self::Item, Self::Item)>
    where
        K: Ord,
        F: FnMut(&Self::Item) -> K;

    fn minmax_by<F>(self, compare: F) -> Option<(Self::Item, Self::Item)>
    where
        F: FnMut(&Self::Item, &Self::Item) -> Ordering;
}

Enriches iterators by adding minmax method.

Required methods

fn minmax(self) -> Option<(Self::Item, Self::Item)>

Gets (min, max).

If the iterator is empty, returns None.

When there are multiple mininum or multiple maximum, returns the first ones as min and max.

Only when the iterator has only one item, it is cloned.

Example

let seq = vec![5, 3, 2, 0, 4, 1];
assert_eq!(seq.into_iter().minmax().unwrap(), (0, 5));

fn minmax_by_key<K, F>(self, key_fn: F) -> Option<(Self::Item, Self::Item)> where
K: Ord,
F: FnMut(&Self::Item) -> K,

Gets (min, max) compared by key_fn.

If the iterator is empty, returns None.

When there are multiple mininum or multiple maximum, returns the first ones as min and max.

Only when the iterator has only one item, it is cloned.

fn minmax_by<F>(self, compare: F) -> Option<(Self::Item, Self::Item)> where
F: FnMut(&Self::Item, &Self::Item) -> Ordering,

Gets (min, max) compared by compare.

If the iterator is empty, returns None.

When there are multiple mininum or multiple maximum, returns the first ones as min and max.

Only when the iterator has only one item, it is cloned.

Loading content...