Struct atcoder_snippets::collections::bitset::BitSet

pub struct BitSet { /* fields omitted */ }

Efficient boolean vector

Methods

impl BitSet

pub fn new(bit_len: usize) -> BitSet

Creates a bitset of the specified bit length.

pub fn len(&self) -> usize

Gets bit length.

pub fn get(&self, i: usize) -> Option<bool>

Gets i-th bit.

If i is out of range, returns None.

pub unsafe fn get_unchecked(&self, i: usize) -> bool

Gets i-th bit without bound checking.

Maybe causes undefined behavior if i is out of range,

pub fn get_mut(&mut self, i: usize) -> Option<BitSetRef>

Gets i-th bit as a mutable boolean.

If i is out of range, returns None.

pub fn at(&mut self, i: usize) -> BitSetRef

Panicking get_mut.

Panic

If i is out of range, it panics.

pub fn count_ones(&self) -> u32

Counts how many bits are sets.

Example

let mut set = BitSet::new(256);
*set.at(0) = true;
*set.at(128) = true;
assert_eq!(set.count_ones(), 2)

Important traits for BitSetBits<'_>
impl<'_> Iterator for BitSetBits<'_> type Item = bool;

pub fn bits(&self) -> BitSetBits

Gets an iterator from the lowest bit to the highest, yielding bools.

Example

let mut set = BitSet::new(4);
*set.at(0) = true;

let bits: Vec<bool> = set.bits().collect();
assert_eq!(bits, vec![true, false, false, false]);

// The iterator is reversible.
let rev_bits: Vec<bool> = set.bits().rev().collect();
assert_eq!(rev_bits, vec![false, false, false, true]);