Module atcoder_snippets::collections::hash_counter

Counter by a hash map from keys to counter values.

AGC031 Aとかが楽に解ける

Structs

HashCounter
Counter by a hash map from keys to counter values.

HashCounterIntoIter
An iterator created by into_iter method on counters.

HashCounterItemRef

HashCounterIter
An iterator created by iter method on counters.

HashCounterValues
An iterator created by values method on counters.