Module atcoder_snippets::collections::heap

Priority queues implemented by binary heaps.

This module covers use cases std::collection::BinaryHeap misses.

You can use both of maximum heap and minimum heap.

You can pass key function for comparing elements. You don't have to implemnt Ord for the element.

As a typical example, std::collection:::BinaryHeap requires a lot of boilerplate for implmeting Dijkstra's single source shortest paths algorithm.

use std::collections::BinaryHeap;
use std::cmp::Ordering;

struct Edge {
    incoming_node: usize,
    cost: u32
}

#[derive(PartialEq, Eq)]
struct Node {
    id: usize,
    distance: u32
}

impl Ord for Node {
    fn cmp(&self, other: &Node) -> Ordering {
        // reversed compare for using a maximum heap as minimum heap
        other.distance.cmp(&self.distance)
    }
}

impl PartialOrd for Node {
    fn partial_cmp(&self, other: &Node) -> Option<Ordering> {
        Some(self.cmp(other))
    }
}

// gets shortest distances for each node from node 0
fn dijkstra(adjacency_list: &[Vec<Edge>]) -> Vec<Option<u32>> {
    let mut queue = BinaryHeap::new();
    let mut distances = vec![None; adjacency_list.len()];
    queue.push(Node { id: 0, distance: 0});
    distances[0] = Some(0);

    while let Some(node) = queue.pop() {
        if node.distance > distances[node.id].unwrap() {
            continue;
        }
        for edge in &adjacency_list[node.id] {
            let next_distance = node.distance + edge.cost;
            if distances[edge.incoming_node].map_or(true, |d| next_distance < d) {
                queue.push(Node { id: edge.incoming_node, distance: next_distance });
                distances[edge.incoming_node] = Some(next_distance);
            }
        }
    }
    distances
}

You can omit impl Ord and impl PartialOrd by using a minimum heap.

struct Edge {
    incoming_node: usize,
    cost: u32
}

struct Node {
    id: usize,
    distance: u32
}

fn dijkstra(adjacency_list: &[Vec<Edge>]) -> Vec<Option<u32>> {
    let mut queue = min_heap_by_key(|node: &Node| node.distance);
    let mut distances = vec![None; adjacency_list.len()];
    queue.push(Node { id: 0, distance: 0});
    distances[0] = Some(0);

    while let Some(node) = queue.pop() {
        if node.distance > distances[node.id].unwrap() {
            continue;
        }
        for edge in &adjacency_list[node.id] {
            let next_distance = node.distance + edge.cost;
            if distances[edge.incoming_node].map_or(true, |d| next_distance < d) {
                queue.push(Node { id: edge.incoming_node, distance: next_distance });
                distances[node.id] = Some(next_distance);
            }
        }
    }
    distances
}

Structs

MaxHeap
Priority queue yielding its maximum item.

MaxHeapIterator

Traits

PriorityQueue
Priority queue trait.

Functions

max_heap
Creates an empty maximum heap.

max_heap_by
Creates an empty maximum heap using cmp for comparison.

max_heap_by_key
Creates an empty maximum heap using key for comparison.

max_heap_from_vec
Creates a maximum heap of all items in vec.

max_heap_from_vec_by
Creates a maximum heap of all items in vec, using cmp for comparison.

max_heap_from_vec_by_key
Creates a maximum heap of all items in vec, using key for comparison.

min_heap
Creates an empty minimum heap.

min_heap_by
Creates an empty minimum heap using cmp for comparison.

min_heap_by_key
Creates an empty minimum heap using key for comparison.

min_heap_from_vec
Creates a minimum heap of all items in vec.

min_heap_from_vec_by
Creates a minimum heap of all items in vec, using cmp for comparison.

min_heap_from_vec_by_key
Creates a minimum heap of all items in vec, using key for comparison.