Trait atcoder_snippets::collections::heap::PriorityQueue

pub trait PriorityQueue<T>: IntoIterator<Item = T> + Extend<T> {
    fn len(&self) -> usize;

    fn is_empty(&self) -> bool;

    fn push(&mut self, x: T);

    fn pop(&mut self) -> Option<T>;

    fn peek(&self) -> Option<&T>;
}

Priority queue trait.

Providing this as a trait is useful for concise coding. When you write a function receiving a MaxHeap, you don't have to write a trait bound for comparison function. If you write type directly, the signature looks like:

use std::cmp::Ordering;

fn f(heap: &mut MaxHeap<i32, impl Fn(&i32, &i32) -> Ordering>) {
    /* ... */
}

Thanks to PriorityQueue trait, you can write as:

fn f(heap: &mut impl PriorityQueue<i32>) {
    /* ... */
}

Required methods

fn len(&self) -> usize

Returns how many items the queue contains.

fn is_empty(&self) -> bool

Returns if the queue is empty.

fn push(&mut self, x: T)

Pushes an item.

fn pop(&mut self) -> Option<T>

Pops the most prioritized item.

fn peek(&self) -> Option<&T>

Gets the most prioritized item without removing it.

Loading content...