Struct atcoder_snippets::collections::vec_union_find_sets::VecUnionFindSets

pub struct VecUnionFindSets { /* fields omitted */ }

Disjoint-set data structure, known as union-find, for integers 0..n.

Thanks to union-by-size and path-compression strategy, average cost of each operation is so much low that it can be regarded as constant time, although theoretically it is not constant.

Methods

impl VecUnionFindSets

pub fn new() -> VecUnionFindSets

Creates empty sets.

pub fn with_items(items_count: usize) -> VecUnionFindSets

Creates count singleton sets.

pub fn add_items(&mut self, count: usize)

Adds count items labeled n..n+count, where n is how many items the sets currently have.

Example

let mut sets = VecUnionFindSets::new();
sets.add_items(3);
assert_eq!(sets.count(), 3);

pub fn items_len(&self) -> usize

Returns how many items are contained by all the sets.

Example

let mut sets = VecUnionFindSets::with_items(2);
assert_eq!(sets.items_len(), 2);
sets.unite(0, 1).unwrap();
assert_eq!(sets.items_len(), 2);

pub fn count(&self) -> usize

Returns how many sets self contains.

Example

let mut sets = VecUnionFindSets::with_items(2);
assert_eq!(sets.count(), 2);
sets.unite(0, 1).unwrap();
assert_eq!(sets.count(), 1);

pub fn len_of(&self, item: usize) -> Result<usize, String>

Returns how many items self contains by the set which has item.

If no set contains item, returns Err with an error message.

Example

let mut sets = VecUnionFindSets::with_items(2);

assert_eq!(sets.len_of(0), Ok(1));
sets.unite(0, 1).unwrap();
assert_eq!(sets.len_of(0), Ok(2));

assert!(sets.len_of(2).is_err());

pub fn set_eq(&self, item1: usize, item2: usize) -> Result<bool, String>

Returns if two sets containing item1 and item2 are the same one.

If no set contains item1 or item2, returns Err with an error message.

Example

let mut sets = VecUnionFindSets::with_items(2);

assert_eq!(sets.set_eq(0, 1), Ok(false));
sets.unite(0, 1);
assert_eq!(sets.set_eq(0, 1), Ok(true));

assert!(sets.set_eq(0, 2).is_err());
assert!(sets.set_eq(2, 3).is_err());

pub fn unite(&mut self, item1: usize, item2: usize) -> Result<bool, String>

Merges two sets, set containing item1 and set containing item2.

If the two sets are same (already merged ones), do nothing and returns Ok(false).

If no set contains item1 or item2, returns Err with an error message.

pub fn iter_cloned(&self) -> impl Iterator<Item = Vec<usize>>

All sets as an iterator yielding Vec<usize>.

Each set is sorted, but the order of sets is unspecified.

Example

use std::collections::HashSet;

let mut sets = VecUnionFindSets::with_items(6);
sets.unite(0, 1).unwrap();
sets.unite(2, 3).unwrap();
sets.unite(3, 4).unwrap();

let sets: HashSet<Vec<usize>> = sets.iter_cloned().collect();
assert_eq!(sets.len(), 3);
assert!(sets.contains(&vec![0, 1]));
assert!(sets.contains(&vec![2, 3, 4]));
assert!(sets.contains(&vec![5]));