Module atcoder_snippets::iter

Enriches iterators.

Structs

Chunks
An iterator created by chunks method on iterators.

GroupBy
An iterator created by group_by method on iterators.

Iterate
An iterator created by iterate function.

IterateMap
An iterator created by iterate_map function.

LScan
An iterator created by lscan method on iterators.

RunLength
An iterator created by run_length method on iterators.

Unfold
An iterator created by unfold function.

Traits

IteratorExt
Enriches iterators by adding various methods.

IteratorInnerProduct
Enriches exact-sized iterators by adding inner_product method.

Functions

iterate
Returns an iterator yielding init, f(init), f(f(init)), ... infinitely.

iterate_map
Returns an iterator applying a function creating a value and a new state infinitely.

unfold
Returns an iterator applying a function creating a value and a new state repeatedly.