Trait atcoder_snippets::iter::IteratorExt

pub trait IteratorExt: Iterator {
    fn chunks(self, size: usize) -> Chunks<Self>
    where
        Self: Sized,
    { ... }

    fn lscan<S, F>(self, state: S, f: F) -> LScan<Self, S, F>
    where
        Self: Sized,
        F: FnMut(&S, Self::Item) -> S,
    { ... }

    fn lscan1<F>(self, f: F) -> Option<LScan<Self, Self::Item, F>>
    where
        Self: Sized,
        Self::Item: Clone,
        F: FnMut(&Self::Item, Self::Item) -> Self::Item,
    { ... }

    fn get_unique(self) -> Option<Self::Item>
    where
        Self: Sized,
        Self::Item: Eq,
    { ... }

    fn group_by<K: Eq, F: FnMut(&Self::Item) -> K>(
        self, 
        f: F
    ) -> GroupBy<K, Self, F>
    where
        Self: Sized,
    { ... }

    fn run_length(self) -> RunLength<Self>
    where
        Self: Sized,
        Self::Item: Eq,
    { ... }

    fn join<T: Display>(self, sep: T) -> String
    where
        Self: Sized,
        Self::Item: Display,
    { ... }

    fn cat(self) -> String
    where
        Self: Sized,
        Self::Item: Display,
    { ... }
}

Enriches iterators by adding various methods.

Provided methods

Important traits for Chunks<I>
impl<I: Iterator> Iterator for Chunks<I> type Item = Vec<I::Item>;

fn chunks(self, size: usize) -> Chunks<Self> where
Self: Sized,

Returns an iterator yielding chunks.

Panic

Panics if size is 0;

Example

let seq = vec![1, 2, 3, 4, 5, 6, 7, 8];
let chunks: Vec<Vec<i32>> = seq.into_iter().chunks(3).collect();
assert_eq!(chunks, vec![vec![1, 2, 3], vec![4, 5, 6], vec![7, 8]]);

Important traits for LScan<I, S, F>
impl<I: Iterator, S, F> Iterator for LScan<I, S, F> where
F: FnMut(&S, I::Item) -> S, type Item = S;

fn lscan<S, F>(self, state: S, f: F) -> LScan<Self, S, F> where
Self: Sized,
F: FnMut(&S, Self::Item) -> S,

Returns an iterator folding the inner iterator and yielding all intermidiate states.

This method acts like Haskell's scanl function. The name lscan corresponds to rfold method of DoubleEndedIterator.

Example

Calculates a cumulative sum.

let data = vec![1, 2, 3, 4, 5];
// the sum of data[start..end] is calculated by cumsum[end] - cumsum[start].
let cumsum: Vec<i32> = data.into_iter().lscan(0, |&acc, x| acc + x).collect();
assert_eq!(cumsum, vec![0, 1, 3, 6, 10, 15]);

fn lscan1<F>(self, f: F) -> Option<LScan<Self, Self::Item, F>> where
Self: Sized,
Self::Item: Clone,
F: FnMut(&Self::Item, Self::Item) -> Self::Item,

lscan using the first item as initial state.

If the iterator is empty, returns None.

Example

Calculates a cumulative max.

use std::cmp;

let data = vec![10, 1, 20, 2, 30];
// maxs[i] means the maximum value in data[..=i].
let maxs: Vec<i32> = data.into_iter()
    .lscan1(|&acc, x| cmp::max(acc, x)).unwrap().collect();
assert_eq!(maxs, vec![10, 10, 20, 20, 30]);

fn get_unique(self) -> Option<Self::Item> where
Self: Sized,
Self::Item: Eq,

If the iterator has any item and all the items are same, returns Some of the first item. Othewise (having no items or non-unique items), returns None.

Important traits for GroupBy<K, I, F>
impl<K: Eq, I: Iterator, F: FnMut(&I::Item) -> K> Iterator for GroupBy<K, I, F> type Item = (K, Vec<I::Item>);

fn group_by<K: Eq, F: FnMut(&Self::Item) -> K>(
self,
f: F
) -> GroupBy<K, Self, F> where
Self: Sized,

fによってグループ分けされたVecを生成するイテレータを返す

SQLのgroup_byではなく、PythonのitertoolやRustのitertoolと同じ意味論である

Example

let seq = vec![('a', 0), ('a', 1), ('a', 2), ('b', 0), ('a', 0), ('a', 1)];
let grouped: Vec<(char, Vec<(char, i32)>)> = seq
    .into_iter()
    .group_by(|&(ch, _)| ch)
    .collect();
assert_eq!(grouped,
           vec![('a', vec![('a', 0), ('a', 1), ('a', 2)]),
                ('b', vec![('b', 0)]),
                ('a', vec![('a', 0), ('a', 1)])]);

Important traits for RunLength<I>
impl<I: Iterator> Iterator for RunLength<I> where
I::Item: Eq, type Item = (I::Item, usize);

fn run_length(self) -> RunLength<Self> where
Self: Sized,
Self::Item: Eq,

Returns an iterator yielding groups of same values as tuples of (value, length).

Example

let seq = vec!['a', 'a', 'a', 'b', 'a', 'a'];
let grouped: Vec<(char, usize)> = seq
    .into_iter()
    .run_length()
    .collect();
assert_eq!(grouped, vec![('a', 3), ('b', 1), ('a', 2)]);

fn join<T: Display>(self, sep: T) -> String where
Self: Sized,
Self::Item: Display,

Concatenates items into a string with interleaving separators.

Example

assert_eq!(&vec![1, 2, 3].into_iter().join(", "), "1, 2, 3");

fn cat(self) -> String where
Self: Sized,
Self::Item: Display,

Concatenates items into a string.

Example

assert_eq!(&vec![1, 2, 3].into_iter().cat(), "123");

Loading content...