Macro atcoder_snippets::read_chunk

macro_rules! read_chunk {
    ( $( $pat:pat = $t:ty ),+ ) => { ... };
}

Reads multiple lines from stdin and create let bindings.

The usage is very similar to read macro, but each binding corresponds to a line, instead of some words.

Example

Solves AtCoder Beginners Selection Practice A - Welcome to AtCoder.

fn main() {
    read_chunk!(
        a = u16,
        (b, c) = (u16, u16),
        s = String
    );
    println!("{} {}", a+b+c, s);
}

The solution can be written using read macro:

fn main() {
    read!(a = u16);
    read!(b = u16, c = u16);
    read!(s = String);
    println!("{} {}", a+b+c, s);
}

but using read_chunk has some advantages:

You don't have to write read several times.

read_chunk gets the mutex for stdin only once. It makes inputting a little bit faster.