Macro atcoder_snippets::readable_collection

macro_rules! readable_collection {
    ($u:ident => $collection_in:ty, $collection_out:ty) => { ... };
    ($u:ident : $( $bound:path ),* => $collection_in:ty, $collection_out:ty) => { ... };
}

Make collection type readable from input line.

The collection type must implement FromIterator.

For example, Vec is readable from input line thanks to the following declaration:

readable_collection!(U => Vec<U>, Vec<U::Output>);

To make HashSet readable, declare as:

readable_collection!(U: Eq, Hash => HashSet<U>, HashSet<U::Output>);

The content of this macro should be either of the followings:

U => collection type for U , collection type for U::Output

U : type bounds of the item type => collection type for U , collection type for U::Output

The first identifier must be U, or the compilation may fail.

Be careful that the separator of type bounds is , not +. This is because of a restriction of Rust's macro system.