Struct atcoder_snippets::modulo::modp::ModP

pub struct ModP { /* fields omitted */ }

A number whose arithmetics is carried modulo a prime number.

Methods

impl ModP

pub unsafe fn set_mod(modulus: ModPModulus) -> Result<(), String>

Sets the modulus.

If modulus is not a prime number, returns Err.

Undefined behaviors

If you make another call of set_mod after creating ModP numbers, you must not use the numbers. The correctness of calculations using the numbers is not guaranteed.

If you call set_mod when two or more threads use ModP numbers, the correctness of calculations using the numbers is not guaranteed.

pub fn new(n: ModPBase) -> ModP

Create a number.

pub unsafe fn new_unchecked(n: ModPBase) -> ModP

Create a number without taking remainder by the modulus.

If n is greater than or equal to the modulus, the correctness of calculations is not guaranteed.

pub fn base(&self) -> ModPBase

Returns a ModPBase satisfying 0 <= x < modulus.

pub fn pow(self, exp: ModPBase) -> ModP

Calculate power using exponentiation by squaring.

Example

unsafe {
    ModP::set_mod(7).unwrap();
}
// 2^5 = 32 = 4 mod 7.
assert_eq!(ModP::new(2).pow(5), ModP::new(4));

pub fn inv(self) -> ModP

Inverse element.

Panic

Panics if self is zero.

Example

// MODULUS is set to 7.
// 3^5 = 15 = 1 mod 7.
unsafe {
    ModP::set_mod(7).unwrap();
}
assert_eq!(ModP::new(3).inv(), ModP::new(5));

pub fn fact_cache() -> FactCache

pub fn inv_cache() -> InvCache

pub fn pow_cache(base: ModPBase) -> PowCache

pub fn combinatorics_cache() -> CombinatoricsCache

Cache for faster calculation.

See CombinatoricsCache.