Function atcoder_snippets::read::read_lines

Important traits for ReadLines<T>
impl<T: ReadableFromLine> Iterator for ReadLines<T> type Item = T::Output;

pub fn read_lines<T: ReadableFromLine>() -> ReadLines<T>

Creates an iterator reading stdin line by line.

Example

Solves AtCoder Beginner Contest 109: Problem B - Shiritori.

use std::collections::HashSet;

// Checks whther the given sequence of words satisfies the constraint.
fn check(words: &[String]) -> bool {
    let mut occurred = HashSet::new();
    let first = &words[0];
    occurred.insert(first);
    let mut last_char = first.chars().last().unwrap();

    for word in &words[1..] {
        if !word.starts_with(last_char) {
            return false;
        }
        if !occurred.insert(word) {
            return false;
        }
        last_char = word.chars().last().unwrap();
    }

    return true;
}

fn main() {
    read!();
    // Uses `yn` snippet.
    yn(check(&read_lines::<String>().collect::<Vec<_>>()));
}

You can read only n lines by take method on the iterator.

// Stdin: "3\n1 10\n2 20\n3 30\n2\n1 10 100\n2 20 200"
read!(n = usize);
let nums1: Vec<(u8, u8)> = read_lines::<(u8, u8)>().take(n).collect();
assert_eq!(nums1, vec![(1, 10), (2, 20), (3, 30)]);

read!();
let nums2: Vec<(u8, u8, u8)> = read_lines::<(u8, u8, u8)>().collect();
assert_eq!(nums2, vec![(1, 10, 100), (2, 20, 200)]);

Deadlock

read_lines gets the mutex for stdin, and release it when the iterator is dropped. So, it causes deadlock to read stdin before the iterator is dropped.