Trait atcoder_snippets::read::Readable

pub trait Readable {
    type Output;

    const WORD_COUNT: usize;

    fn read_words(words: &[&str]) -> Result<Self::Output, String>;
}

Readable from a constant number of words.

A word is a fragment of an input line splitted by whiltespaces.

The following types are readable from one word:

Unit type ()

char

String

All primitibe numeric types (isize, usize, i8, u8, i16, u16, i32, u32, i64, u64, f32, f64)

A tuple of an arbitrary number of Readables is Readable.

readモジュールは、1-origin整数を読み込んで0-origin整数にする操作を、 特殊なReadable型を提供することによって可能としている。 cf. usize_

To make a custom readable type, use readable macro instead of implementing this trait directly.

Example

See implementation of Vec2.

Associated Types

type Output

Output type.

Usually Output should be Self, but there are some exceptions:

1-origin to 0-origin conversion by usize_[struct.usize_.html] etc.

Chars

Loading content...
Associated Constants

const WORD_COUNT: usize

Returns how many words are read.

Loading content...
Required methods

fn read_words(words: &[&str]) -> Result<Self::Output, String>

Converts words into Outputs.

If the conversion fails, returns Err with error message.

Panics

If words.len() differs from WORD_COUNT, the method may panic.

Loading content...