Trait atcoder_snippets::slice::SliceExt

pub trait SliceExt<T> {
    fn fill(&mut self, val: T)
    where
        T: Clone;

    fn span<F: Fn(&T) -> bool>(&self, predicate: F) -> (&[T], &[T]);

    fn group_by<K: Eq, F: Fn(&T) -> K>(
        &self, 
        key_fn: F
    ) -> SliceGroupBy<T, K, F>;

    fn split_by_gap<F: Fn(&T, &T) -> bool>(&self, gap_fn: F) -> SplitByGap<T, F>;

    fn permutations(&self) -> Permutations<T>;

    fn count_inversions(&self) -> usize
    where
        T: Clone + Ord;
}

Enriches slices by adding various methods.

Required methods

fn fill(&mut self, val: T) where
T: Clone,

Fills the slice by val.

Example

let mut seq = vec![0; 5];
seq[1..4].fill(1);
assert_eq!(seq, vec![0, 1, 1, 1, 0]);

fn span<F: Fn(&T) -> bool>(&self, predicate: F) -> (&[T], &[T])

Takes the prefix slice satisfying the predicate, and makes a tuple of the slice and the rest.

Example

let seq = [1, 2, 3, 4, 1, 2];
assert_eq!(seq.span(|&x| x < 3), ([1, 2].as_ref(), [3, 4, 1, 2].as_ref()));
assert_eq!(seq.span(|&x| x < 9), ([1, 2, 3, 4, 1, 2].as_ref(), [].as_ref()));
assert_eq!(seq.span(|&x| x < 1), ([].as_ref(), [1, 2, 3, 4, 1, 2].as_ref()));

Important traits for SliceGroupBy<'a, T, K, F>
impl<'a, T, K: Eq, F: Fn(&T) -> K> Iterator for SliceGroupBy<'a, T, K, F> type Item = (K, &'a [T]);

fn group_by<K: Eq, F: Fn(&T) -> K>(&self, key_fn: F) -> SliceGroupBy<T, K, F>

Returns an iterator yielding groups.

Each group is a pair of key and subslice.

Example

let seq = [('a', 0), ('a', 1), ('a', 2), ('b', 0), ('a', 0), ('a', 1)];
let grouped: Vec<(char, Vec<(char, i32)>)> = seq
    .group_by(|&(ch, _)| ch)
    .map(|(ch, pairs)| (ch, pairs.to_vec()))
    .collect();
assert_eq!(grouped,
           vec![('a', vec![('a', 0), ('a', 1), ('a', 2)]),
                ('b', vec![('b', 0)]),
                ('a', vec![('a', 0), ('a', 1)])]);

Important traits for SplitByGap<'a, T, F>
impl<'a, T, F: Fn(&T, &T) -> bool> Iterator for SplitByGap<'a, T, F> type Item = &'a [T];

fn split_by_gap<F: Fn(&T, &T) -> bool>(&self, gap_fn: F) -> SplitByGap<T, F>

Returns an iterator yielding subslices separated by gap_fn.

gap_fn takes 2 items. When gap_fn(&self[i-1], &self[i]) is true, self is splitted at i.

Example

let seq = [1, 2, 3, 1, 2, 1, 1];
// splits seq into increasing subsequences
let mut subseqs = seq.split_by_gap(|&a, &b| a > b);
assert_eq!(subseqs.next(), Some(vec![1, 2, 3].as_slice()));
assert_eq!(subseqs.next(), Some(vec![1, 2].as_slice()));
assert_eq!(subseqs.next(), Some(vec![1, 1].as_slice()));
assert_eq!(subseqs.next(), None);

Important traits for Permutations<'a, T>
impl<'a, T: 'a> Iterator for Permutations<'a, T> type Item = Vec<&'a T>;

fn permutations(&self) -> Permutations<T>

Returns an iterator yielding all permutations of the slice.

Each permutation is a Vec of shared references to items in the slice.

Example

let mut perms = [1, 2, 3].permutations();
assert_eq!(perms.next(), Some(vec![&1, &2, &3]));
assert_eq!(perms.next(), Some(vec![&1, &3, &2]));
assert_eq!(perms.next(), Some(vec![&2, &1, &3]));
assert_eq!(perms.next(), Some(vec![&2, &3, &1]));
assert_eq!(perms.next(), Some(vec![&3, &1, &2]));
assert_eq!(perms.next(), Some(vec![&3, &2, &1]));
assert_eq!(perms.next(), None);

fn count_inversions(&self) -> usize where
T: Clone + Ord,

Counts the number of pairs of indices (i, j) satisfing i < j and self[i] > self[j].

Example

assert_eq!([1, 0, 3, 2, 0].count_inversions(), 5);

Loading content...