Struct atcoder_snippets::table::Table

pub struct Table<T> { /* fields omitted */ }

2-dimentional array.

In competitive programming, a lot of problems require handling 2-dimentional tables. This struct provides utility methods for a table.

All rows in a table must have the same length.

Table can have 0-length rows, but it's impossible to have 0-length columns.

Methods

impl<T> Table<T>

pub fn from_rows(rows: Vec<Vec<T>>) -> Option<Table<T>>

Creates a new table from rows represented as Vec<Vec<T>>.

All rows must have the same length. Otherwise, returns None.

pub unsafe fn from_rows_unchecked(rows: Vec<Vec<T>>) -> Table<T>

Create a new table from rows without shape checking.

pub fn height(&self) -> usize

Number of rows.

pub fn width(&self) -> usize

Number of columns.

pub fn shape(&self) -> (usize, usize)

(height, column).

pub fn indices(&self) -> impl Iterator<Item = (usize, usize)>

Creates an iterator yielding table indices as (y, x) in the dictionary order.

Example

let table = table![0; 2,3];
let indices: Vec<(usize, usize)> = table.indices().collect();
assert_eq!(indices, vec![
    (0, 0), (0, 1), (0, 2),
    (1, 0), (1, 1), (1, 2)
]);

pub fn inside(&self, (y, x): (usize, usize)) -> bool

Checks the index is in range.

Example

let table = table![0; 5,5];
assert_eq!(table.inside((4, 4)), true);
assert_eq!(table.inside((4, 5)), false);

pub fn row(&self, y: usize) -> Option<&[T]>

Row as a slice.

pub fn row_mut(&mut self, y: usize) -> Option<&mut [T]>

Row as a mutable slice.

pub fn get(&self, (y, x): (usize, usize)) -> Option<&T>

Gets the shared reference to an items at (y, x).

pub fn get_mut(&mut self, (y, x): (usize, usize)) -> Option<&mut T>

Gets the mutable reference to an items at (y, x).

Important traits for TableRows<'a, T>
impl<'a, T> Iterator for TableRows<'a, T> type Item = &'a [T];

pub fn rows(&self) -> TableRows<T>

pub fn map<U>(self, f: impl FnMut(T) -> U) -> Table<U>

Maps each element by f and gets a new table with the same shape.

Example

let rows = vec![
    vec![1, 2, 3],
    vec![10, 20, 30]
];
let table = Table::from_rows(rows).unwrap();

let mapped_rows = vec![
    vec![2, 4, 6],
    vec![20, 40, 60]
];
let mapped_table = Table::from_rows(mapped_rows).unwrap();

assert_eq!(table.map(|x| x*2), mapped_table);

pub fn ref_map<U>(&self, f: impl FnMut(&T) -> U) -> Table<U>

Same as map, but takes self by reference instead by move.

Example

let rows = vec![
    vec![1, 2, 3],
    vec![10, 20, 30]
];
let table = Table::from_rows(rows).unwrap();

let mapped_rows_1 = vec![
    vec![2, 4, 6],
    vec![20, 40, 60]
];
let mapped_table_1 = Table::from_rows(mapped_rows_1).unwrap();
assert_eq!(table.ref_map(|&x| x*2), mapped_table_1);

// You can use table again.
let mapped_rows_2 = vec![
    vec![3, 6, 9],
    vec![30, 60, 90]
];
let mapped_table_2 = Table::from_rows(mapped_rows_2).unwrap();
assert_eq!(table.ref_map(|&x| x*3,), mapped_table_2);

pub fn transpose(&mut self)

Transposes the table's row and column.

Example

let rows = vec![
    vec![1, 2, 3],
    vec![10, 20, 30]
];
let mut table = Table::from_rows(rows).unwrap();

let transposed_rows = vec![
    vec![1, 10],
    vec![2, 20],
    vec![3, 30],
];
let transposed_table = Table::from_rows(transposed_rows).unwrap();

table.transpose();
assert_eq!(table, transposed_table);

When you do same operations on rows and columns, transpose is very useful for short coding.

// fn operation_on_rows(table: &Table<i32>) {
//     ...
// }
//
// let mut table = ...;

for _ in 0..2 {
    operation_on_rows(&table);
    table.transpose();
}

pub fn adjacent4_indices(
&self,
(y, x): (usize, usize)
) -> Option<Vec<(usize, usize)>>

Indices of vertically and horizontlly adjacent cells in the dictionary order.

Useful for traversing a grid graph.

If the given index is out of range, returns None.

Example

let table = table![0; 5,5];
assert_eq!(table.adjacent4_indices((3,2)),
           Some(vec![(2,2), (3,1), (3,3), (4,2)]));
assert_eq!(table.adjacent4_indices((4,3)),
           Some(vec![(3,3), (4,2), (4,4)]));

pub fn adjacent8_indices(
&self,
(y, x): (usize, usize)
) -> Option<Vec<(usize, usize)>>

Indices of 8 enclosing cells in the dictionary order.

Useful for traversing a grid graph.

If the given index is out of range, returns None.

Example

let table = table![0; 5,5];
assert_eq!(table.adjacent8_indices((3,2)),
           Some(vec![(2,1), (2,2), (2,3), (3,1), (3,3), (4,1), (4,2), (4,3)]));
assert_eq!(table.adjacent8_indices((4,3)),
           Some(vec![(3,2), (3,3), (3,4), (4,2), (4,4)]));

pub fn accumulate<F1, F2>(
&self,
identity: T,
op: F1,
op_inv: F2
) -> CumulativeTable<T, F1, F2> where
T: Clone,
F1: Fn(&T, &T) -> T,
F2: Fn(&T, &T) -> T,

Creates a cumulative table that can handle 2-dimentional range sum queries, etc.

Example

let rows = vec![
    vec![0,    1,    2,    3,    4],
    vec![0,   10,   20,   30,   40],
    vec![0,  100,  200,  300,  400],
    vec![0, 1000, 2000, 3000, 4000],
];
let table = Table::from_rows(rows).unwrap();
let cumulative = table.accumulate(0, |&a,&b| a+b, |&a,&b| a-b);
// Calculates the sum of this subtable:
// [[ 20,  30,  40],
//  [200, 300, 400]]
assert_eq!(cumulative.query_yx(1..=2, 2..).unwrap(), 990)