Module atcoder_snippets::z

Exact string matching by Z algorithm.

Z algorithm takes advantage of Z-array. Z-array of string S is a sequence of integers with the same length as S satisfing the following condition: for all i (0 ≤ i < length of S), S[i] is the maximum l such that S[..l] matches S[i..i+l], with one exception that S[0] = 0 (in case 0 < i).

Exact string matching is performed by the following steps:

For the given text string T and pattern string P, constructs S, which is a concatenation of P, $ and T. $ is a character contained by neither T nor P.

Gets the Z-array of S.

Find all elements in the Z-array whose value is the length of P.

For example, let T be aaabbbabbaabb and P be aabb. Then S and the Z-array is constructed as follows:

      S: a a b b $ a a a b b b a b b a a b b
Z-array: 0 1 0 0 0 2 4 1 0 0 0 1 0 0 4 1 0 0

So the algorithm find 2 occurrences of P in T.

Structs

LongestPrefixLengths
An iterator generated by longest_prefix_lengths.

ZMatchIndices
An iterator generated by z_match_indices.

Traits

ZString
A string (sequence) equipped with methods for Z-array and exact string matching.