Trait atcoder_snippets::z::ZString

pub trait ZString<T: Eq + Clone> {
    fn longest_prefix_lengths(&self) -> LongestPrefixLengths<T>;

    fn z_match_indices(&self, pattern: &Self) -> ZMatchIndices<Option<T>>;
}

A string (sequence) equipped with methods for Z-array and exact string matching.

Required methods

Important traits for LongestPrefixLengths<'a, T>
impl<'a, T: Eq> Iterator for LongestPrefixLengths<'a, T> type Item = usize;

fn longest_prefix_lengths(&self) -> LongestPrefixLengths<T>

Gets an iterator yielding Z-array's values.

Definition of Z-array is stated in the module documentation. The iterator yields all Z-array's values except for the 1st one (always 0).

Exahausting the iterator takes Θ(self.len()) time.

Example

let text: Vec<char> = "abababaab".chars().collect();
let mut indices = text.longest_prefix_lengths();

assert_eq!(indices.next(), Some(0)); // match length for "abababaab" and "bababaab"
assert_eq!(indices.next(), Some(5)); // match length for "abababaab" and "ababaab"
assert_eq!(indices.next(), Some(0)); // match length for "abababaab" and "babaab"
assert_eq!(indices.next(), Some(3)); // match length for "abababaab" and "abaab"
assert_eq!(indices.next(), Some(0)); // match length for "abababaab" and "baab"
assert_eq!(indices.next(), Some(1)); // match length for "abababaab" and "aab"
assert_eq!(indices.next(), Some(2)); // match length for "abababaab" and "ab"
assert_eq!(indices.next(), Some(0)); // match length for "abababaab" and "b"
assert_eq!(indices.next(), None);

In some problem, you will takes advantage of this preprosessing rather than exact string matching.

For example, to solve Codeforces Round #578 (Div. 2), Problem E - Compress Words, you have to find the length of the longest string, which is both S's suffix and T's prefix, in O(length of T) time. This can be implemented using longest_prefix_length.

use std::cmp;

fn length(s: &str, t: &str) -> usize {
    let s: Vec<char> = s.chars().collect();
    let t: Vec<char> = t.chars().collect();

    let len = cmp::min(s.len(), t.len());
    let s_trimmed = &s[s.len() - len..];
    let t_trimmed = &t[..len];

    // Let s_trimmed be "sample" and t_trimmed be "please".
    // Then concat becomes "pleasesample".
    let concat: Vec<char> = t_trimmed.iter().cloned()
        .chain(s_trimmed.iter().cloned())
        .collect();

    // The last 6 elements of the Z-array is [0, 0, 0, 3, 0, 0].
    // The length to be returned is 3
    // because it's the 3rd element of the Z-array counted from the end of the array,
    // so first 3 characters in t_trimmed match the last 3 characters in s_trimmed.
    concat.longest_prefix_lengths().skip(len-1).enumerate()
        .find(|&(i, l)| len - i == l)
        .map(|(_, l)| l)
        .unwrap_or(0)
}

assert_eq!(length("sample", "please"), 3);
assert_eq!(length("samplease", "ease"), 4);
assert_eq!(length("samplease", "in"), 0);
assert_eq!(length("sampleasein", "out"), 0);

Important traits for ZMatchIndices<T>
impl<T: Eq + Clone> Iterator for ZMatchIndices<T> type Item = usize;

fn z_match_indices(&self, pattern: &Self) -> ZMatchIndices<Option<T>>

Gets an iterator yielding self's indices matching pattern.

Exahausting the iterator takes Θ(self.len() + pattern.len()) time.

Example

let text: Vec<char> = "xababxabababxabxabab".chars().collect();
let pattern: Vec<char> = "abab".chars().collect();
let mut indices = text.z_match_indices(&pattern);

// "xababxabababxabxabab" matches "abab" at position 1, 6, 8 and 16.
assert_eq!(indices.next(), Some(1));
assert_eq!(indices.next(), Some(6));
assert_eq!(indices.next(), Some(8));
assert_eq!(indices.next(), Some(16));
assert_eq!(indices.next(), None);

Loading content...